requirements.txt
gunicorn~=22.0.0
parse~=1.20.1
pytest~=8.2.0
pytest-cov~=5.0.0
requests-wsgi-adapter~=0.4.1
WebOb~=1.8.7
